pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.14",
    "discord-py>=2.5.2",
    "flask>=3.1.1",
    "python-dotenv>=1.1.1",
    "pytz>=2025.2",
    "pillow>=10.3.0"
]

